# Performance work on `vcpkg search` (30x improvement)