--- a/include/vcpkg/registries.h
+++ b/include/vcpkg/registries.h
@@ -120,7 +120,7 @@ namespace vcpkg
         void set_default_registry(std::unique_ptr<RegistryImplementation>&& r);
         void set_default_registry(std::nullptr_t r);
         bool is_default_builtin_registry() const;
-        void set_default_builtin_registry_baseline(StringView baseline) const;
+        void set_default_builtin_registry_baseline(StringView baseline);
 
         // returns whether the registry set has any modifications to the default
         // (i.e., whether `default_registry` was set, or `registries` had any entries)