… src/vcpkg/paragraphs.cpp
… src/vcpkg/registries.cpp
… src/vcpkg/vcpkgpaths.cpp