Results: `vcpkg search zlib` is 30x faster in my Linux docker container. The biggest sources of improvement were:
1. Eliminate uses of `std::regex`
2. Avoid loading and parsing the manifests multiple times
3. Optimize low level utf8 reader

I've also left in basic performance counters to enable future performance tracking and improvement.

Before perf work:
```
$ time vcpkg search zlib --debug
...
[DEBUG] Time in subprocesses: 0 us
[DEBUG] Time in parsing JSON: 89751 us
[DEBUG] Time in JSON reader: 743202 us
[DEBUG] Time in filesystem: 12453 us
[DEBUG] Time in loading ports: 862885 us
[DEBUG] Exiting after 876.1 ms (876052 us)

real    0m0.878s
user    0m0.848s
sys     0m0.030s
```
After perf work:
```
$ time vcpkg search zlib --debug
...
[DEBUG] Time in subprocesses: 0 us
[DEBUG] Time in parsing JSON: 8732 us
[DEBUG] Time in JSON reader: 6051 us
[DEBUG] Time in filesystem: 4993 us
[DEBUG] Time in loading ports: 23459 us
[DEBUG] Exiting after 28.84 ms (28809 us)

real    0m0.031s
user    0m0.017s
sys     0m0.009s
```

More numbers from Windows (Surface Book 3): 4.5x faster

From hot without perf improvements:
```
[DEBUG] Time in parsing JSON: 189672 us
[DEBUG] Time in JSON reader: 294580 us
[DEBUG] Time in filesystem: 813409 us
[DEBUG] Time in loading ports: 1731497 us
[DEBUG] Exiting after 1.898 s (1837987 us)
```

From cold with perf improvements:
```
[DEBUG] Time in parsing JSON: 46451 us
[DEBUG] Time in JSON reader: 44898 us
[DEBUG] Time in filesystem: 9744531 us
[DEBUG] Time in loading ports: 9921944 us
[DEBUG] Exiting after 10.08 s (9947811 us)
```

hot with perf improvements:
```
[DEBUG] Time in parsing JSON: 30963 us
[DEBUG] Time in JSON reader: 21867 us
[DEBUG] Time in filesystem: 300172 us
[DEBUG] Time in loading ports: 401880 us
[DEBUG] Exiting after 418.1 ms (416803 us)
```